+++ views/layouts/main.handlebars
@@ -24,12 +24,12 @@
             <!-- if user logged in, render this profile and logout, otherwise render login and sign-up -->
             {{#if logged_in}}
             <li><a href="/profile" class="text-white">Profile</a></li>
-            <li><a href="#" class="text-white">Logout</a></li>
+            <li><a href="/logout" class="text-white">Logout</a></li>
             {{else}}
             <li><a href="/login" class="text-white">Login</a></li>
             <li><a href="/signup" class="text-white">Sign Up</a></li>
             {{/if}}
-            <li><a href="#" class="text-white">About</a></li>
+            <li><a href="" class="text-white">About</a></li>
           </ul>
         </div>
       </div>
@@ -53,7 +53,7 @@
 
   <!-- Render script for logged in users only -->
   {{#if logged_in}}
-  <script src="js/logout.js"></script>
+  {{!-- <script src="js/logout.js"></script> --}}
   {{/if}}
 <script src="/js/bootstrap.bundle.min.js"></script>
 </body>
