+++ views/login.handlebars
@@ -28,19 +28,19 @@
             <input
               type='email'
               class='form-control rounded-3'
-              id='floatingInput'
+              id='email-login'
               placeholder='name@example.com'
             />
-            <label for='floatingInput'>Email address</label>
+            <label for='email-login'>Email address</label>
           </div>
           <div class='form-floating mb-3'>
             <input
               type='password'
               class='form-control rounded-3'
-              id='floatingPassword'
+              id='password-login'
               placeholder='Password'
             />
-            <label for='floatingPassword'>Password</label>
+            <label for='password-login'>Password</label>
           </div>
           <button
             id="login-btn"
